refactor: split OAuth callback server
